# backend/openapi.json
{"openapi":"3.0.2","info":{"title":"Tercer Parcial - API REST. Ingeniería Web","description":"José Luis Bueno Pachón.","version":"1.0"},"paths":{"/lineas":{"get":{"tags":["Lineas"],"summary":"Get Lineas","operationId":"get_lineas","parameters":[{"required":false,"schema":{"title":"Numlinea","type":"integer"},"name":"numLinea","in":"query"},{"required":false,"schema":{"title":"Sentido","type":"integer"},"name":"sentido","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Lineas","type":"array","items":{"$ref":"#/components/schemas/Linea"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/lineas/{nombre}":{"get":{"tags":["Lineas"],"summary":"Get Lineas By Nombre","operationId":"get_lineas_by_nombre","parameters":[{"required":true,"schema":{"title":"Nombre","type":"string"},"name":"nombre","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Lineas By Nombre","type":"array","items":{"$ref":"#/components/schemas/Linea"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/lineas/{latitud}/{longitud}":{"get":{"tags":["Lineas"],"summary":"Get Lineas By Latitud Longitud","operationId":"get_lineas_by_latitud_longitud_lineas__latitud___longitud__get","parameters":[{"required":true,"schema":{"title":"Latitud","type":"number"},"name":"latitud","in":"path"},{"required":true,"schema":{"title":"Longitud","type":"number"},"name":"longitud","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/viviendas":{"get":{"tags":["Ignorar"],"summary":"Get Houses","operationId":"get_houses","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Houses","type":"array","items":{"$ref":"#/components/schemas/Vivienda"}}}}}}},"post":{"tags":["Ignorar"],"summary":"Create House","operationId":"new_house","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewVivienda"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vivienda"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/viviendas/{id}":{"get":{"tags":["Ignorar"],"summary":"Get House","operationId":"get_house_by_id","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vivienda"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Ignorar"],"summary":"Edit House","operationId":"edit_house","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditVivienda"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vivienda"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Ignorar"],"summary":"Delete House","operationId":"delete_house","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth":{"get":{"tags":["Auth"],"summary":"Auth","operationId":"auth_auth_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/usuarios/{id}":{"get":{"tags":["Ignorar"],"summary":"Get User","operationId":"get_user_by_id","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Ignorar"],"summary":"Edit User","operationId":"edit_user","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditUsuario"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Ignorar"],"summary":"Delete User","operationId":"delete_user","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/usuarios":{"get":{"tags":["Ignorar"],"summary":"Get User By Filter","operationId":"get_user_by_filter","parameters":[{"required":false,"schema":{"title":"Nombre","type":"string"},"name":"nombre","in":"query"},{"required":false,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get User By Filter","type":"array","items":{"$ref":"#/components/schemas/Usuario"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Ignorar"],"summary":"Create User","operationId":"new_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUsuario"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EditUsuario":{"title":"EditUsuario","required":["nombre","email","apellidos","direccion","latitude","longitude","foto"],"type":"object","properties":{"nombre":{"title":"Nombre","type":"string"},"email":{"title":"Email","type":"string"},"apellidos":{"title":"Apellidos","type":"string"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"foto":{"title":"Foto","type":"string"}}},"EditVivienda":{"title":"EditVivienda","required":["titulo","descripcion","precio","direccion","latitude","longitude"],"type":"object","properties":{"titulo":{"title":"Titulo","type":"string"},"descripcion":{"title":"Descripcion","type":"string"},"precio":{"title":"Precio","exclusiveMinimum":0.0,"type":"number"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"images":{"title":"Images","type":"array","items":{"type":"string"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Linea":{"title":"Linea","required":["codLinea","nombreLinea","sentido","orden","codParada","nombreParada","direccion","lat","lon"],"type":"object","properties":{"_id":{"title":" Id","type":"string"},"codLinea":{"title":"Codlinea","type":"integer"},"nombreLinea":{"title":"Nombrelinea","type":"string"},"sentido":{"title":"Sentido","type":"integer"},"orden":{"title":"Orden","type":"string"},"codParada":{"title":"Codparada","type":"string"},"nombreParada":{"title":"Nombreparada","type":"string"},"direccion":{"title":"Direccion","type":"string"},"lat":{"title":"Lat","type":"number"},"lon":{"title":"Lon","type":"number"}}},"NewUsuario":{"title":"NewUsuario","required":["nombre","email","apellidos","direccion","latitude","longitude","foto"],"type":"object","properties":{"nombre":{"title":"Nombre","type":"string"},"email":{"title":"Email","type":"string"},"apellidos":{"title":"Apellidos","type":"string"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"foto":{"title":"Foto","type":"string"}}},"NewVivienda":{"title":"NewVivienda","required":["titulo","descripcion","user","precio","direccion","latitude","longitude"],"type":"object","properties":{"titulo":{"title":"Titulo","type":"string"},"descripcion":{"title":"Descripcion","type":"string"},"user":{"title":"User","type":"string"},"precio":{"title":"Precio","type":"number"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"images":{"title":"Images","type":"array","items":{"type":"string"}}}},"Usuario":{"title":"Usuario","required":["nombre","email","apellidos","direccion","latitude","longitude","foto"],"type":"object","properties":{"_id":{"title":" Id","type":"string"},"nombre":{"title":"Nombre","type":"string"},"email":{"title":"Email","type":"string"},"apellidos":{"title":"Apellidos","type":"string"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"foto":{"title":"Foto","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"Vivienda":{"title":"Vivienda","required":["user","titulo","descripcion","precio","direccion","latitude","longitude"],"type":"object","properties":{"_id":{"title":" Id","type":"string"},"user":{"title":"User","type":"string"},"titulo":{"title":"Titulo","type":"string"},"descripcion":{"title":"Descripcion","type":"string"},"precio":{"title":"Precio","type":"number"},"direccion":{"title":"Direccion","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"images":{"title":"Images","type":"array","items":{"type":"string"}}}}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}